kernel: interrupt_controller to library
