+++ kernel/devices/ps2/src/CMakeLists.txt
@@ -12,4 +12,5 @@ add_library(libps2 STATIC
 )
 
 target_include_directories(libps2 PUBLIC ../include)
-target_link_libraries(libps2 PUBLIC libkernelcxx)
+target_link_libraries(libps2 PUBLIC libkernelcxx
+	PRIVATE interrupt_controller)
