+++ kernel/x86/include/x86/gdt.hpp
@@ -10,6 +10,31 @@ namespace x86 {
 void reload_segment_registers();
 void init_task_register();
 
+
+enum class SegmentIndex: int
+{
+	null = 0,
+	kernel_code = 1,
+	kernel_data = 2,
+	single_task = 3,
+};
+
+constexpr inline
+uint16_t segment_selector(SegmentIndex index, TableIndicator ti,
+			  ProtectionRing required_privilege_level) noexcept
+{
+	return segment_selector(static_cast<uint16_t>(index), ti, required_privilege_level);
+}
+
+/// Kernel code segment.
+constexpr uint16_t KERNEL_CODE_SELECTOR =
+	segment_selector(SegmentIndex::kernel_code, TableIndicator::GLOBAL, ProtectionRing::kernel);
+
+/// Kernel data segment.
+constexpr uint16_t KERNEL_DATA_SELECTOR =
+	segment_selector(SegmentIndex::kernel_data, TableIndicator::GLOBAL, ProtectionRing::kernel);
+
+
 class GDT {
 public:
 	GDT();
